+++ forui/test/src/widgets/slider/slider_golden_test.dart
@@ -171,6 +171,9 @@ void main() {
       }
 
       testWidgets('${theme.name} - $layout - focused', (tester) async {
+        FocusManager.instance.highlightStrategy = .alwaysTraditional;
+        addTearDown(() => FocusManager.instance.highlightStrategy = .automatic);
+
         await tester.pumpWidget(
           TestScaffold.app(
             theme: theme.data,